Introduce a method to delay worker shutdown.
